requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aerich==0.7.2
    # via -r requirements.in
aiosqlite==0.17.0
    # via tortoise-orm
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via -r requirements.in
bcrypt==5.0.0
    # via passlib
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
click==8.3.0
    # via
    #   aerich
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   uvicorn
coverage==7.11.0
    # via pytest-cov
cryptography==46.0.3
    # via python-jose
dictdiffer==0.9.0
    # via aerich
ecdsa==0.19.1
    # via python-jose
fastapi==0.115.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.27.0
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
iniconfig==2.3.0
    # via pytest
iso8601==1.1.0
    # via tortoise-orm
packaging==25.0
    # via pytest
passlib==1.7.4
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.23
    # via cffi
pydantic==2.12.3
    # via
    #   aerich
    #   fastapi
    #   pydantic-settings
    #   tortoise-orm
pydantic-core==2.41.4
    # via pydantic
pydantic-settings==2.5.0
    # via -r requirements.in
email-validator>=2.0.0
    # via -r requirements.in
pypika-tortoise==0.1.6
    # via tortoise-orm
pytest==8.3.0
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==0.24.0
    # via -r requirements.in
pytest-cov==5.0.0
    # via -r requirements.in
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   uvicorn
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.9
    # via -r requirements.in
pytz==2025.2
    # via tortoise-orm
pyyaml==6.0.3
    # via uvicorn
rsa==4.9.1
    # via python-jose
    # via -r requirements.in
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.38.6
    # via fastapi
tomlkit==0.13.3
    # via aerich
tortoise-orm==0.21.0
    # via
    #   -r requirements.in
    #   aerich
typing-extensions==4.15.0
    # via
    #   aiosqlite
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uvicorn==0.30.0
    # via -r requirements.in
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
